Move KPF exceptions out of vital into arrows
