+++ arrows/kpf/yaml/kpf_parse_utils.cxx
@@ -38,7 +38,7 @@
 
 #include "kpf_parse_utils.h"
 
-#include <vital/exceptions/kpf.h>
+#include <arrows/kpf/yaml/kpf_exception.h>
 
 #include <utility>
 #include <cctype>
